refactor(topics): address review — split classifier, extract taxonomy, fix CLI help
- Extract the frozen 20-label taxonomy into topic-taxonomy.ts, dropping the
  classifier module to focused logic-only code.
- Split anchor preparation from row scoring via createTopicClassifier: the
  label anchors are embedded once (not once per profile artifact), and
  anchor-embed failure is now a distinct failure boundary from per-row
  inference failure.
- Fix the corrupted analyse/query usage text: topic-candidates and top-sites
  each get their own self-contained option block.
- Delete the dead sha256File(modelPath) load (hashed ~30 MB then discarded).
- Drop the unused profile_path/commit_state columns from the topic-candidates
  query SELECT/QueryRow.
